# internal/openapi/spec.yaml
openapi: 3.1.0
info:
  title: System Wrangler API
  version: v1
  description: |
    HTTP API exposed by the System Wrangler backend. Every protected
    endpoint requires the `sw_session` cookie issued by
    `POST /api/auth/login` (or `POST /api/auth/setup` on a fresh
    install). State-changing requests must also carry the
    `X-Sw-Csrf: 1` header — the CSRF middleware rejects authenticated
    POST/PUT/PATCH/DELETE without it.

    This spec is the source of truth for the API surface. A drift test
    (`cmd/server/openapi_drift_test.go`) compares it to the routes
    registered on the production mux so the two cannot rot apart.
servers:
  - url: /
    description: Same-origin (the SPA and the API share the host).
...
security:
  - sessionCookie: []
paths:
...
  /api/systems/{id}/group:
    put:
      tags: [groups]
      summary: Set or clear a system's group membership.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId]
              properties:
                groupId:
                  type: [string, "null"]
                  description: Target group, or `null` to clear.
      responses:
        "204": { description: Membership updated. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/groups:
    get:
      tags: [groups]
      summary: List groups visible to the caller.
      responses:
        "200":
          description: Groups the caller can see.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [groups]
      summary: Create a group. Global Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created. `Location` carries `/api/groups/{id}`.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/groups/{id}:
    get:
      tags: [groups]
      summary: One group by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: The matching group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Not found, or hidden from the caller's scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [groups]
      summary: Rename a group. Global Admin only.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "200":
          description: Renamed group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [groups]
      summary: Delete a group. Systems become ungrouped.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Group deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/backup:
    post:
      tags: [backup]
      summary: Stream a SQLite snapshot of the database. Global Admin only.
      responses:
        "200":
          description: The snapshot file streamed back with a timestamped filename.
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/vnd.sqlite3:
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Another backup is already in progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sw_session
  responses:
    BadRequest:
      description: Request payload rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but the scope forbids this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or is hidden from the caller's scope).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicts with existing state (duplicate or precondition failed).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Group:
      type: object
      required: [id, name, createdAt, systemCount]
      properties:
        id: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        systemCount: { type: integer }
    GroupInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
...